requirements.txt
Pillow
pyobjc
pyautogui
pytesseract